requirements.txt
streamlit
yfinance
pandas
numpy
pandas_ta
requests
beautifulsoup4
pyyaml
plotly
schedule
